Fill with data

+++ src/main/java/com/ahold/allwaysdelivery/ride/payload/Ride.java
@@ -45,7 +45,7 @@ public class Ride {
     @Field
     private boolean bonus;
     @Field
-    private Integer fee;
+    private Double fee;
     @Field
     private RideReservationStatus status = RideReservationStatus.PENDING;
 
@@ -97,11 +97,11 @@ public class Ride {
         this.bonus = bonus;
     }
 
-    public Integer getFee() {
+    public Double getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Double fee) {
         this.fee = fee;
     }
